// app/shaga/src/generated/types/AffairPayload.ts
/**
 * This code was GENERATED using the solita package.
 * Please DO NOT EDIT THIS FILE, instead rerun solita to update it or write a wrapper to add functionality.
 *
 * See: https://github.com/metaplex-foundation/solita
 */

import * as beet from '@metaplex-foundation/beet'
import { HashAlgorithm, hashAlgorithmBeet } from './HashAlgorithm'
export type AffairPayload = {
  coordinates: string
  ipAddress: string
  cpuName: string
  gpuName: string
  totalRamMb: number
  solPerHour: beet.bignum
  affairTerminationTime: beet.bignum
  hashAlgorithm: HashAlgorithm
  privatePairHash: beet.COption<Uint8Array>
}

/**
 * @category userTypes
 * @category generated
 */
export const affairPayloadBeet = new beet.FixableBeetArgsStruct<AffairPayload>(
  [
    ['coordinates', beet.utf8String],
    ['ipAddress', beet.utf8String],
    ['cpuName', beet.utf8String],
    ['gpuName', beet.utf8String],
    ['totalRamMb', beet.u32],
    ['solPerHour', beet.u64],
    ['affairTerminationTime', beet.u64],
    ['hashAlgorithm', hashAlgorithmBeet],
    ['privatePairHash', beet.coption(beet.bytes)],
  ],
  'AffairPayload'
)