// app/shaga/src/generated/types/HashAlgorithm.ts
/**
 * This code was GENERATED using the solita package.
 * Please DO NOT EDIT THIS FILE, instead rerun solita to update it or write a wrapper to add functionality.
 *
 * See: https://github.com/metaplex-foundation/solita
 */

import * as beet from '@metaplex-foundation/beet'
/**
 * @category enums
 * @category generated
 */
export enum HashAlgorithm {
  None,
  Sha256,
  Keccak,
  Blake3,
}

/**
 * @category userTypes
 * @category generated
 */
export const hashAlgorithmBeet = beet.fixedScalarEnum(
  HashAlgorithm
) as beet.FixedSizeBeet<HashAlgorithm, HashAlgorithm>